adding a new activity : ExistingList
so to show all the lists added so far, we use this activity. Activity
will query the db using the custom cmd string, to get the first two
columns. which will be later supplied to the list adapter . List adapter
is a custom adapter but more customizations to come, like checkbox
within each row for example.  Problems to solve : How to map each ID ->
ITEM in a listview ?